requirements.txt
# Core dependencies
duckdb>=1.0.0
pyyaml>=6.0
streamlit>=1.30.0
pandas>=2.0.0
boto3>=1.34.0

# Configuration & Validation
pydantic>=2.0.0
python-dotenv>=1.0.0

# Retry & Resilience
tenacity>=8.2.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code Quality
black>=24.0.0
ruff>=0.1.0
mypy>=1.7.0